+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 9.0.0
 
 * Add home location to collectionobject schema and advanced search
+* Denormalize media priority list
 * Add password complexity requirements to tenant bindings
 * Add password validators which verify passwords meet requirements set by tenants
 * Enable strict checksum verification so that builds fail on corrupted Maven artifact downloads